feature(ch24): tweaks to support dependent project builds (#23)
* chore: added a build watch script, and xlink for linking dependencies
* feature: consolidated sass variables into a standalone file
  * Many of our sass variables will need to be referenced and used in dependent
projects for consistency. I extracted all variable definitions to a standalone
_variables.scss file that can be not only used in our internal modules, but
also independently imported into dependent projects for direct reference. The
best example of this is variables that define a common collection of viewport
breakpoints for use in CSS media queries, but color variables will also be
useful as well.
* feature: added a couple more utility functions to the util package
* feature: added options for list layout using either grid or flex